backend/src/requirements.txt
asgiref==3.2.3
cffi==1.14.0
cryptography==2.8
Django==3.0.3
django-cors-headers==3.2.1
django-rest-knox==4.1.0
django-versatileimagefield==2.0
djangocms-admin-style==1.5.0
djangorestframework==3.11.0
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2==2.8.4
pycparser==2.19
python-magic==0.4.15
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
